# src/test/resources/openapi/approved-openapi.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "technical-interview-demo API",
    "description": "Machine-readable contract for the demo application's supported external /api/** surface. Internal-only overview, documentation, OpenAPI publication, and actuator validation paths are intentionally excluded.",
    "version": "APPROVED"
  },
...
  "paths": {
...
    "/api/admin/users/{id}/roles": {
      "put": {
        "tags": [
          "Admin Users"
        ],
        "summary": "Replace managed user roles",
        "description": "Requires an authenticated session with the ADMIN role, a valid same-site CSRF header mirrored from the readable XSRF-TOKEN cookie, and replaces the non-bootstrap role set for one persisted user.",
        "operationId": "replaceRoles",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          },
          {
            "name": "X-XSRF-TOKEN",
            "in": "header",
            "description": "Same-site CSRF header whose value must match the readable XSRF-TOKEN cookie.",
            "required": true
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AdminUserRoleUpdateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminUserAccountResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request payload.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiProblemResponse"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid authenticated session.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiProblemResponse"
                }
              }
            }
          },
          "403": {
            "description": "Authenticated user does not have the ADMIN role.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiProblemResponse"
                }
              }
            }
          },
          "404": {
            "description": "Persisted user was not found.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiProblemResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "sessionCookie": []
          }
        ]
      }
    },
    "/api/account/language": {
      "put": {
        "tags": [
          "Account"
        ],
        "summary": "Update the account language",
        "description": "Requires an authenticated session and a valid same-site CSRF header mirrored from the readable XSRF-TOKEN cookie. Blank or null clears the preference.",
        "operationId": "updatePreferredLanguage",
        "parameters": [
          {
            "name": "X-XSRF-TOKEN",
            "in": "header",
            "description": "Same-site CSRF header whose value must match the readable XSRF-TOKEN cookie.",
            "required": true
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserAccountLanguageRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/UserAccountResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "sessionCookie": []
          }
        ]
      }
    },
...
    "/api/admin/users": {
      "get": {
        "tags": [
          "Admin Users"
        ],
        "summary": "List persisted users",
        "description": "Requires an authenticated session with the ADMIN role and returns persisted users together with current roles and role-grant provenance.",
        "operationId": "listUsers",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/AdminUserAccountResponse"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid authenticated session.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiProblemResponse"
                }
              }
            }
          },
          "403": {
            "description": "Authenticated user does not have the ADMIN role.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiProblemResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "sessionCookie": []
          }
        ]
      }
    },
...
    "/api/account": {
      "get": {
        "tags": [
          "Account"
        ],
        "summary": "Get the current account",
        "description": "Requires an authenticated session and returns the persisted application account.",
        "operationId": "currentUser",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/UserAccountResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "sessionCookie": []
          }
        ]
      }
    }
  },
  "components": {
    "schemas": {
...
      "ApiProblemResponse": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string",
            "description": "Problem title.",
            "example": "Unauthorized"
          },
          "status": {
            "type": "integer",
            "description": "HTTP status code.",
            "format": "int32",
            "example": 401
          },
          "detail": {
            "type": "string",
            "description": "Technical problem detail kept stable for debugging and logs.",
            "example": "Authentication is required to access this resource."
          },
          "messageKey": {
            "type": "string",
            "description": "Stable localization key for the error type.",
            "example": "error.request.unauthorized"
          },
          "message": {
            "type": "string",
            "description": "Localized end-user message resolved from the request language.",
            "example": "You must authenticate before performing this operation."
          },
          "language": {
            "type": "string",
            "description": "Two-letter ISO 639-1 language code actually used for the localized message.",
            "example": "en"
          }
        },
        "description": "Localized ProblemDetail payload returned for API errors."
      },
...
      "AdminUserRoleUpdateRequest": {
        "required": [
          "reason",
          "roles"
        ],
        "type": "object",
        "properties": {
          "roles": {
            "minItems": 1,
            "uniqueItems": true,
            "type": "array",
            "description": "Replacement role set. `USER` must always be present.",
            "items": {
              "type": "string",
              "description": "Role name to keep assigned after replacement.",
              "example": "ADMIN",
              "enum": [
                "USER",
                "ADMIN"
              ]
            }
          },
          "reason": {
            "maxLength": 255,
            "minLength": 0,
            "type": "string",
            "description": "Short operator-supplied explanation for the role change.",
            "example": "Needs audit review access."
          }
        },
        "description": "Payload for replacing the managed role set of one persisted user."
      },
      "AdminUserAccountResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "description": "Database identifier.",
            "format": "int64",
            "example": 1
          },
          "provider": {
            "type": "string",
            "description": "OAuth provider name.",
            "example": "github"
          },
          "login": {
            "type": "string",
            "description": "External login from the authenticated OAuth provider.",
            "example": "kamkie"
          },
          "displayName": {
            "type": "string",
            "description": "Display name from the OAuth provider.",
            "example": "Kamil Kiewisz"
          },
          "email": {
            "type": "string",
            "description": "Email address when available from the provider."
          },
          "preferredLanguage": {
            "type": "string",
            "description": "Preferred two-letter language code used as the last localization fallback."
          },
          "roles": {
            "type": "array",
            "description": "Current roles assigned to the persisted user.",
            "items": {
              "type": "string",
              "description": "Granted role name.",
              "example": "USER"
            }
          },
          "roleGrants": {
            "type": "array",
            "description": "Persisted role-grant provenance entries.",
            "items": {
              "$ref": "#/components/schemas/AdminUserRoleGrantResponse"
            }
          },
          "lastLoginAt": {
            "type": "string",
            "description": "UTC instant of the latest authenticated request.",
            "format": "date-time"
          },
          "createdAt": {
            "type": "string",
            "description": "Creation timestamp as a UTC instant.",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "description": "Last update timestamp as a UTC instant.",
            "format": "date-time"
          }
        },
        "description": "Persisted application user profile plus role-grant provenance for ADMIN review."
      },
      "AdminUserRoleGrantResponse": {
        "type": "object",
        "properties": {
          "role": {
            "type": "string",
            "description": "Granted role name.",
            "example": "ADMIN"
          },
          "source": {
            "type": "string",
            "description": "How the role grant entered application state.",
            "example": "BOOTSTRAP",
            "enum": [
              "AUTHENTICATED_LOGIN",
              "BOOTSTRAP",
              "ADMIN_MANAGED"
            ]
          },
          "grantedAt": {
            "type": "string",
            "description": "UTC instant when this role grant was recorded.",
            "format": "date-time"
          },
          "grantedByUserId": {
            "type": "integer",
            "description": "Granting application user id when the role was assigned manually.",
            "format": "int64",
            "example": 1
          },
          "grantedByLogin": {
            "type": "string",
            "description": "Granting application user login when the role was assigned manually.",
            "example": "admin-user"
          },
          "reason": {
            "type": "string",
            "description": "Optional operator-supplied reason captured for manually managed grants."
          }
        },
        "description": "Persisted provenance for one granted application role."
      },
      "UserAccountLanguageRequest": {
        "type": "object",
        "properties": {
          "preferredLanguage": {
            "type": "string",
            "description": "Two-letter ISO 639-1 language code. Use null or blank to clear the preference.",
            "example": "pl"
          }
        },
        "description": "Payload for updating the current user's preferred language."
      },
      "UserAccountResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "description": "Database identifier.",
            "format": "int64",
            "example": 1
          },
          "provider": {
            "type": "string",
            "description": "OAuth provider name.",
            "example": "github"
          },
          "login": {
            "type": "string",
            "description": "External login from the authenticated OAuth provider.",
            "example": "kamkie"
          },
          "displayName": {
            "type": "string",
            "description": "Display name from the OAuth provider.",
            "example": "Kamil Kiewisz"
          },
          "email": {
            "type": "string",
            "description": "Email address when available from the provider."
          },
          "preferredLanguage": {
            "type": "string",
            "description": "Preferred two-letter language code used as the last localization fallback."
          },
          "roles": {
            "type": "array",
            "description": "Roles assigned to the persisted user.",
            "items": {
              "type": "string",
              "description": "Granted role name.",
              "example": "USER"
            }
          },
          "lastLoginAt": {
            "type": "string",
            "description": "UTC instant of the latest authenticated request.",
            "format": "date-time"
          },
          "createdAt": {
            "type": "string",
            "description": "Creation timestamp as a UTC instant.",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "description": "Last update timestamp as a UTC instant.",
            "format": "date-time"
          }
        },
        "description": "Persisted application profile for the current authenticated user."
      },
...
    },
    "securitySchemes": {
      "sessionCookie": {
        "type": "apiKey",
        "description": "Authenticated browser session cookie used by protected operations. It is established through a configured identity provider login path under /api/session/oauth2/authorization/{registrationId} when the optional oauth profile is active. Unsafe browser writes also require the X-XSRF-TOKEN request header mirrored from the readable XSRF-TOKEN cookie.",
        "name": "technical-interview-demo-session",
        "in": "cookie"
      }
    }
  }
}